fix merchant fee calculation decimal

+++ apac-services-exchange-order-module/apac-services-exchange-order-service/src/main/java/com/cwt/bpg/cbt/exchange/order/calculator/HkAirCalculator.java
@@ -168,7 +168,7 @@ public class HkAirCalculator implements Calculator<AirFeesBreakdown, AirFeesInpu
                     mFTotal = mFTotal.add(transactionFee);
                 }
             }
-            merchantFeeAmount = BigDecimal.ZERO.max(round(calculatePercentage(mFTotal.doubleValue(), merchantFee.getMerchantFeePercent()), scale, roundingMode));
+            merchantFeeAmount = BigDecimal.ZERO.max(round(calculatePercentage(mFTotal, merchantFee.getMerchantFeePercent()), scale, roundingMode));
         }
 
         return merchantFeeAmount;
